USFM Importer Round-trip fix
Fixed the missing metadata information for the importer/exporter to work.

Got rid of the old unused USFM importer files that were messing with the structure of our file system, and the repo is now cleaner.

The importer was tested with several files both imported and exported correctly through Round-trip and USFM generate.
