+++ webviews/codex-webviews/src/NewSourceUploader/importers/usfm/index.tsx
@@ -2,10 +2,10 @@ import { ImporterPlugin } from "../../types/plugin";
 import { BookOpen } from "lucide-react";
 import { UsfmImporterForm } from "./UsfmImporterForm";
 
-export const usfmImporterPlugin: ImporterPlugin = {
-    id: "usfm",
+export const usfmExperimentalImporterPlugin: ImporterPlugin = {
+    id: "usfm-experimental",
     name: "USFM Biblical Texts",
-    description: "Import Unified Standard Format Marker files for biblical texts",
+    description: "Import Unified Standard Format Marker files for biblical texts with round-trip export support",
     icon: BookOpen,
     component: UsfmImporterForm,
     supportedExtensions: ["usfm", "sfm", "SFM", "USFM"],
