--- webviews/codex-webviews/src/NewSourceUploader/importers/usfm/types.ts
@@ -1,2 +1,0 @@
-// Deprecated: previous external grammar types removed. This file intentionally left minimal.
-export { };
